+++ spade-typeinference/src/requirements.rs
@@ -91,7 +91,7 @@ impl Requirement {
 
     /// Check if there are updates that allow us to resolve the requirement.
     /// - If target_type is still `Unknown`, we don't know how to resolve the requirement
-    /// - Otherwise it will either be unsatsifieable. i.e. the new type does not fulfil the
+    /// - Otherwise it will either be unsatisfiable. i.e. the new type does not fulfill the
     /// requirement, in which case an error is returned
     /// - Or the requirement is now satisfied, in which case new unification tasks which are
     /// applied due to the result are returned. After this, the constraint is no longer needed